Refactor and add warning
