+++ sdk/python/tests/compiler/testdata/condition.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env python3
-# Copyright 2019 Google LLC
+# Copyright 2018 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,29 +12,43 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import kfp
-from kfp import dsl
 
-def print_op(msg):
-    """Print a message."""
-    return dsl.ContainerOp(
-        name='Print',
-        image='alpine:3.6',
-        command=['echo', msg],
-    )
+import kfp.dsl as dsl
+
+
+class FlipCoinOp(dsl.ContainerOp):
+
+  def __init__(self, name):
+    super(FlipCoinOp, self).__init__(
+      name=name,
+      image='python:alpine3.6',
+      command=['sh', '-c'],
+      arguments=['python -c "import random; result = \'heads\' if random.randint(0,1) == 0 '
+                 'else \'tails\'; print(result)" | tee /tmp/output'],
+      file_outputs={'output': '/tmp/output'})
+
+
+class PrintOp(dsl.ContainerOp):
+  
+  def __init__(self, name, msg):
+    super(PrintOp, self).__init__(
+      name=name,
+      image='alpine:3.6',
+      command=['echo', msg])
     
 
 @dsl.pipeline(
-    name='Conditional Example Pipeline',
-    description='Shows how to use dsl.Condition().'
+  name='pipeline flip coin',
+  description='shows how to use dsl.Condition.'
 )
-def conditional_pipeline(num: int = 5):
+def flipcoin():
+  flip = FlipCoinOp('flip')
 
-    with dsl.Condition(num == 5):
-        print_op('Number is equal to 5')
-    with dsl.Condition(num != 5):
-        print_op('Number is not equal to 5')
+  with dsl.Condition(flip.output=='heads'):
+    flip2 = FlipCoinOp('flip-again')
 
+    with dsl.Condition(flip2.output=='tails'):
+      PrintOp('print1', flip2.output)
 
-if __name__ == '__main__':
-    kfp.compiler.Compiler().compile(conditional_pipeline, __file__ + '.yaml')
+  with dsl.Condition(flip.output=='tails'):
+      PrintOp('print2', flip2.output)
