+++ sdk/python/tests/compiler/testdata/condition.yaml
@@ -6,14 +6,15 @@ spec:
   check:
     args:
     - "EXITCODE=$(python -c 'import sys\ninput1=str.rstrip(sys.argv[1])\ninput2=str.rstrip(sys.argv[2])\n\
-      try:\n  input1=int(input1)\n  input2=int(input2)\nexcept Error:\n  input1=str(input1)\n\
-      print(0) if (input1 == input2) else print(1)' $(params.num) 5); exit $EXITCODE"
+      try:\n  input1=int(input1)\n  input2=int(input2)\nexcept:\n  input1=str(input1)\n\
+      print(0) if (input1 == input2) else print(1)' '$(params.flip)' 'heads'); exit\
+      \ $EXITCODE"
     command:
     - sh
     - -c
     image: python:alpine3.6
   params:
-  - name: num
+  - name: flip
 ---
 apiVersion: tekton.dev/v1alpha1
 kind: Condition
@@ -23,67 +24,144 @@ spec:
   check:
     args:
     - "EXITCODE=$(python -c 'import sys\ninput1=str.rstrip(sys.argv[1])\ninput2=str.rstrip(sys.argv[2])\n\
-      try:\n  input1=int(input1)\n  input2=int(input2)\nexcept Error:\n  input1=str(input1)\n\
-      print(0) if (input1 != input2) else print(1)' $(params.num) 5); exit $EXITCODE"
+      try:\n  input1=int(input1)\n  input2=int(input2)\nexcept:\n  input1=str(input1)\n\
+      print(0) if (input1 == input2) else print(1)' '$(params.flip-again)' 'tails');\
+      \ exit $EXITCODE"
     command:
     - sh
     - -c
     image: python:alpine3.6
   params:
-  - name: num
+  - name: flip-again
+---
+apiVersion: tekton.dev/v1alpha1
+kind: Condition
+metadata:
+  name: condition-3
+spec:
+  check:
+    args:
+    - "EXITCODE=$(python -c 'import sys\ninput1=str.rstrip(sys.argv[1])\ninput2=str.rstrip(sys.argv[2])\n\
+      try:\n  input1=int(input1)\n  input2=int(input2)\nexcept:\n  input1=str(input1)\n\
+      print(0) if (input1 == input2) else print(1)' '$(params.flip)' 'tails'); exit\
+      \ $EXITCODE"
+    command:
+    - sh
+    - -c
+    image: python:alpine3.6
+  params:
+  - name: flip
 ---
 apiVersion: tekton.dev/v1beta1
 kind: Task
 metadata:
-  name: print
+  name: flip
 spec:
+  results:
+  - description: /tmp/output
+    name: output
   steps:
-  - command:
-    - echo
-    - Number is equal to 5
-    image: alpine:3.6
-    name: print
+  - args:
+    - python -c "import random; result = 'heads' if random.randint(0,1) == 0 else
+      'tails'; print(result)" | tee $(results.output.path)
+    command:
+    - sh
+    - -c
+    image: python:alpine3.6
+    name: flip
 ---
 apiVersion: tekton.dev/v1beta1
 kind: Task
 metadata:
-  name: print-2
+  name: flip-again
 spec:
+  results:
+  - description: /tmp/output
+    name: output
+  steps:
+  - args:
+    - python -c "import random; result = 'heads' if random.randint(0,1) == 0 else
+      'tails'; print(result)" | tee $(results.output.path)
+    command:
+    - sh
+    - -c
+    image: python:alpine3.6
+    name: flip-again
+---
+apiVersion: tekton.dev/v1beta1
+kind: Task
+metadata:
+  name: print1
+spec:
+  params:
+  - name: flip-again-output
   steps:
   - command:
     - echo
-    - Number is not equal to 5
+    - $(inputs.params.flip-again-output)
     image: alpine:3.6
-    name: print-2
+    name: print1
+---
+apiVersion: tekton.dev/v1beta1
+kind: Task
+metadata:
+  name: print2
+spec:
+  params:
+  - name: flip-again-output
+  steps:
+  - command:
+    - echo
+    - $(inputs.params.flip-again-output)
+    image: alpine:3.6
+    name: print2
 ---
 apiVersion: tekton.dev/v1beta1
 kind: Pipeline
 metadata:
   annotations:
-    pipelines.kubeflow.org/pipeline_spec: '{"description": "Shows how to use dsl.Condition().",
-      "inputs": [{"default": "5", "name": "num", "optional": true, "type": "Integer"}],
-      "name": "Conditional Example Pipeline"}'
-  name: conditional-example-pipeline
+    pipelines.kubeflow.org/pipeline_spec: '{"description": "shows how to use dsl.Condition.",
+      "name": "pipeline flip coin"}'
+  name: pipeline-flip-coin
 spec:
-  params:
-  - default: '5'
-    name: num
+  params: []
   tasks:
+  - name: flip
+    params: []
+    taskRef:
+      name: flip
   - conditions:
     - conditionRef: condition-1
       params:
-      - name: num
-        value: $(params.num)
-    name: print
+      - name: flip
+        value: $(tasks.flip.results.output)
+    name: flip-again
     params: []
     taskRef:
-      name: print
+      name: flip-again
   - conditions:
     - conditionRef: condition-2
       params:
-      - name: num
-        value: $(params.num)
-    name: print-2
-    params: []
+      - name: flip-again
+        value: $(tasks.flip-again.results.output)
+    - conditionRef: condition-1
+      params:
+      - name: flip
+        value: $(tasks.flip.results.output)
+    name: print1
+    params:
+    - name: flip-again-output
+      value: $(tasks.flip-again.results.output)
     taskRef:
-      name: print-2
+      name: print1
+  - conditions:
+    - conditionRef: condition-3
+      params:
+      - name: flip
+        value: $(tasks.flip.results.output)
+    name: print2
+    params:
+    - name: flip-again-output
+      value: $(tasks.flip-again.results.output)
+    taskRef:
+      name: print2
